refactor(providers): extract internal_api_providers tests to sibling file
AGENTS.rust.md §5: the file was 2868 lines with >50% tests. All 8 inline
test modules move verbatim to internal_api_providers_tests.rs behind the
standard #[cfg(test)] #[path] shim; source drops to 1276 lines.

Pure move, zero logic change: source lines 1-1271 and the moved test body
are byte-identical to the previous revision. right: 452 passed, 6 skipped
before and after.
